lint

+++ tests/meltano/api/controllers/test_settings.py
@@ -5,8 +5,6 @@ from flask import url_for
 class TestRoles:
     def test_delete_admin_role(self, api, app):
         with app.test_request_context():
-            res = api.delete(url_for("settings.roles"), json={
-                "role": "admin",
-            })
+            res = api.delete(url_for("settings.roles"), json={"role": "admin"})
 
         assert res.status_code == 403
